node app now gets config data from django
Settings are stored in swag/settings.py, but this will probably be replaced by a more managable solution later.

Also note that this relies on having either an xapi lrs or something that proxies to an xapi lrs at '/xapi'. For example,
this can be done in nginx by adding something like the following to the default site:

    location /xapi/ {
        proxy_pass http://staging.tunapanda.org/learninglocker/public/data/xAPI/;
    }